将String转为long逻辑改为先转为decimal，防止报错

+++ common/src/main/java/com/alibaba/datax/common/element/LongColumn.java
@@ -96,7 +96,7 @@ public class LongColumn extends Column {
 			return null;
 		}
 
-		return (double) (long) this.asLong();
+		return this.asBigDecimal().doubleValue();
 	}
 
 	@Override
